feat: 딜러 구현
