+++ README.md
@@ -17,10 +17,10 @@
 
 ### 딜러
 
-- [ ] 딜러는 두 장의 카드를 받는다.
-- [ ] 카드 숫자 합이 16이하인지 검사한다.
-    - [ ] 반드시 카드를 추가로 받는다.
-- [ ] 딜러는 본인 결과를 알고 있다.
+- [x] 딜러는 두 장의 카드를 받는다.
+- [x] 카드 숫자 합이 16이하인지 검사한다.
+    - [x] 반드시 카드를 추가로 받는다.
+- [x] 딜러는 본인 결과를 알고 있다.
 
 ### 카드
