+++ src/main/java/domain/Dealer.java
@@ -1,0 +1,11 @@
+package domain;
+
+public class Dealer extends Player {
+    public Dealer(final Name name, final Cards cards) {
+        super(name, cards);
+    }
+
+    public boolean isSumUnderStandard() {
+        return sumOfPlayerCards() < 17;
+    }
+}
